Move controllers to auth namespace
