+++ app/controllers/auth/confirmations_controller.rb
@@ -1,0 +1,5 @@
+module Auth
+  class ConfirmationsController < Devise::ConfirmationsController
+    def required; end
+  end
+end
