+++ app/controllers/auth/sessions_controller.rb
@@ -1,0 +1,5 @@
+module Auth
+  class SessionsController < Devise::SessionsController
+    skip_before_action :authenticate_user!
+  end
+end
