--- app/controllers/confirmations_controller.rb
@@ -1,3 +1,0 @@
-class ConfirmationsController < Devise::ConfirmationsController
-  def required; end
-end
