--- app/controllers/sessions_controller.rb
@@ -1,3 +1,0 @@
-class SessionsController < Devise::SessionsController
-  skip_before_action :authenticate_user!
-end
